Drop lucide-react, move ThemeProvider + pages off the root barrel (0.5.0)
Dep-hygiene + page-demotion half of the titan 0.5.0 release (plan A2). Three
breaking changes, so they land together for 0.5.0. Version stays 0.4.0 here —
documented, not bumped.

- Remove lucide-react entirely. Replace the three Workout icon uses
  (WeightBadge Dumbbell, PrBadge / PrHistoryModal Star) with inline SVGs
  (src/.../Workout/icons.tsx) that mirror lucide's exact glyph paths, viewBox,
  stroke, fill, and round line caps — rendering is unchanged. Same raw-<svg>
  web-SVG mechanism Progress.tsx already uses (lucide-react was web-only too).
  Also drops the now-vestigial lucide-react-native optional peer.

- Move the ThemeProvider VALUE off the root barrel onto the /theme subpath.
  ThemeProvider imports nativewind (vars()); re-exporting it from root dragged
  the nativewind chain into every consumer of @titan-design/react-ui. The root
  now re-exports a nativewind-free theme surface (src/theme/barrel.ts); theme
  utilities (getSemanticColors, resolveColor, shadows, elevation, presets, …)
  stay on root; ThemeProvider types stay on root (erased at build).

- Demote the page-level organisms (ActiveWorkoutPage, ExerciseDetailPage,
  ProgramPlanningPage) to a new /pages subpath, mirroring /bodymap: value
  exports move to src/pages.ts, type-only re-exports stay on the root barrel.
  TrainingStatusPage stays on /bodymap (body-highlighter dep).

- Disable tsup code-splitting so entries are self-contained. Splitting had
  merged the nativewind-importing ThemeProvider into a theme-utils chunk the
  root entry also imported, re-leaking nativewind into root's ESM graph even
  after it left the barrel. Self-contained entries keep root nativewind-free
  (as the CJS output already was).

- Add Workout/README.md tier map (atoms/molecules/organisms/pages) plus the
  /bodymap + /pages rationale and the BaseBadge internal-primitive exemption.
